sgs-project/server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-socketio>=5.8.0
pydantic>=2.0.0
websockets>=11.0
redis>=5.0.0
aiofiles>=23.0.0
httpx>=0.24.0

sqlmodel>=0.0.14
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6